Gmail specific archive + delete treatment
After trial and error, archiving and deleting needs a different
treatment than other email providers. Gmail is detected using regex
match on the Maildir name: either "Gmail" or "Google". Adapt as needed.

Additionally, IMAP settings in Gmail should be:

- When I mark a message in IMAP as deleted: Auto-Expunge off - Wait for
the client to update the server.
- When a message is marked as deleted and expunged from the last visible
IMAP folder: Move the message to the Trash