Better handling for aliases
Determining context based on the "TO" field instead of where the email
landed on disk. Better for email composing using aliases.

It should also help with "refile" and "trash" which also use context
resolution (via the `mu4e--get-folder` function) and default to the
current context which could lead to refiling in the wrong mailbox.